Fix potential buffer overflow
The makefile included in 1999-fall/mp2 assembles the code by calling
MASM.EXE.  When building with MASM 6.14.8444, MASM.EXE is a
compatibility driver which invokes ML.EXE with the /Zm option.  The /Zm
option enables MASM 5.10 compatibility.  When this compatibility mode is
enabled, the following instruction assembles without error:

  mov     [di], '$'

However, MASM assumes the above instruction is equivalent to:

  mov     DWORD PTR [di], '$'

Which would lead to a buffer overflow when di equals MAX_BUFF_LEN - 1.

Explicitly specify that the register indirect memory operand is of type
BYTE to correct the potential buffer overflow.

Note: The actual release of MASM 5.10 reports an error when attempting
to assemble the above line.  It is unclear why MASM 6.14.8444 allows
this construct in MASM 5.10 compatibility mode.