Fix Parse procedure to accept number after ')'
The Parse procedure fails to accept a number after a close parentheses.
For example:

Please enter an equation.
Hit <ESC>, or enter an empty equation to exit.
(2)3
Verifying and parsing the string...
Type 6 error: Encountered a number while expecting a binary operator.
The equation was not fully simplified due to an error.

Fix the Parse procedure such that the above expression does not
generated an error.