When '-' prefixes digits, encode a negative number
When a '-' prefixes a string of digits in the input buffer, use the
ascbin procedure to encode the substring as a negative number in the
control string.  This change makes the Parse procedure more closely
conform with the reference implementation.

Previously, the Parse procedure placed a negate operator in the control
string followed by the positive number.  This resulted in more steps
when solving the equation.