Fix ASCII to binary number conversion
The ascbin procedure can sometimes return a status code of 6 which is
undocumented in the ECE 291 lab manual.  Status code 6 appears to be
returned in the following instances:

  - The last numerical digit is immediately followed by a '+', '-', or
    '.'.
  - The first character in the number is a '.'.
  - There are no numerical digits, just two characters which can be
    either '+', '-', or '.'.

In the case that the last numerical digit is immediately followed by a
'+' or '-' operator, the binary number conversion is always positive.
In the case that ascbin returns status code 6, check to see if the first
character was a '-' and if so, negate the binary value.