Fix < operator to avoid potential spurious errors
The shr and shl instructions can sometimes result in the carry flag
being set which would be interpreted as an error flag.  Clear the carry
flag before returning.  Explicitly clearing the carry flag before
returning is not necessary on all code paths as their execution depends
on the carry flag being unset.